A bug was found when a project attempted to write a non-printable character into the log.
This bug does not present itself until you attempt to save the log file to Xml as these characters might be considered valid if outputting to the console; therefore it does not occur until the very end of the build.

The fix is to sanitize all of the messages attempting to be written to the log using the new Utilities.SanitizeMessageForXml(string); there has been no perf testing on this.

Added unit tests (NUnit3) around some of the areas affected.